+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - For English, changed `Expert`'s difficulty level rename to `Turn Me Undead`. For obvious reasons.
 - Removed all unused and functionless legacy content from compilation: `Charcoal`, `Calendars`, `Gramophones`, `Stone Axes`, `Stone Pickaxes`, `Ash Clots`, `Walking Ashes`. These bits will be reintroduced when required.
 - Wall flips and wall rolls were made slightly more reliable, with their code rewritten.
+- Grappling hooks' code was made more data-driven, with some vanilla hook range values corrected.
 ### Configuration
 - Existing features received the following new options:
   **Guns:** `EnableMinigunDynamicFirerate`, `EnableStarCannonDynamicFirerate`.